BAEL-2520 code moved to java-jpa module (#6525)
* BAEL-2520 code changes done

* BAEL-2520 moved code to java-jpa module

* BAEL-2520 moved code to java-jpa module
