+++ persistence-modules/java-jpa/src/main/java/com/baeldung/jpa/criteria/entity/Item.java
@@ -1,0 +1,49 @@
+package com.baeldung.jpa.criteria.entity;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Table(name = "item")
+@Entity
+public class Item {
+
+    @Id
+    private Long id;
+    private String color;
+    private String grade;
+    private String name;
+
+    public String getColor() {
+        return color;
+    }
+
+    public String getGrade() {
+        return grade;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setColor(String color) {
+        this.color = color;
+    }
+
+    public void setGrade(String grade) {
+        this.grade = grade;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
