+++ persistence-modules/java-jpa/src/main/java/com/baeldung/jpa/criteria/repository/CustomItemRepository.java
@@ -1,0 +1,12 @@
+package com.baeldung.jpa.criteria.repository;
+
+import java.util.List;
+
+import com.baeldung.jpa.criteria.entity.Item;
+
+public interface CustomItemRepository {
+
+    List<Item> findItemsByColorAndGrade();
+
+    List<Item> findItemByColorOrGrade();
+}
