+++ persistence-modules/java-jpa/src/main/resources/META-INF/persistence.xml
@@ -87,5 +87,26 @@
             <property name="eclipselink.logging.parameters" value="true"/>
         </properties>
     </persistence-unit>
+    
+    <persistence-unit name="jpa-h2-criteria">
+		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
+		<class>com.baeldung.jpa.criteria.entity.Item</class>
+		<exclude-unlisted-classes>true</exclude-unlisted-classes>
+		<properties>
+			<property name="javax.persistence.jdbc.driver"
+				value="org.h2.Driver" />
+			<property name="javax.persistence.jdbc.url"
+				value="jdbc:h2:mem:test" />
+			<property name="javax.persistence.jdbc.user" value="sa" />
+			<property name="javax.persistence.jdbc.password" value="" />
+			<property name="hibernate.dialect"
+				value="org.hibernate.dialect.H2Dialect" />
+			<property name="hibernate.hbm2ddl.auto" value="create-drop" />
+			<property name="show_sql" value="true" />
+			<property name="hibernate.temp.use_jdbc_metadata_defaults"
+				value="false" />
+			<property name="javax.persistence.sql-load-script-source" value="item.sql"/>
+		</properties>
+	</persistence-unit>
 
 </persistence>
