+++ persistence-modules/java-jpa/src/test/resources/item.sql
@@ -1,0 +1,4 @@
+insert into item(id,grade,color) values (10,'C','blue');
+insert into item(id,grade,color) values (11,'C','red');
+insert into item(id,grade,color) values (12,'A','blue');
+insert into item(id,grade,color) values (13,'D','red');
