+++ persistence-modules/spring-data-jpa/src/main/java/com/baeldung/dao/repositories/CustomItemRepository.java
@@ -1,7 +1,5 @@
 package com.baeldung.dao.repositories;
 
-import java.util.List;
-
 import org.springframework.stereotype.Repository;
 
 import com.baeldung.domain.Item;
@@ -15,7 +13,4 @@ public interface CustomItemRepository {
 
     void findThenDelete(Long id);
 
-    List<Item> findItemsByColorAndGrade();
-
-    List<Item> findItemByColorOrGrade();
 }
